Change import statements
Change import statements and so also change calls of the imported modules.

Do this to make this software work again.
The cause for that this software stopped working is not clear and understood to me.

Describe when projects using this software stopped working because of this software:
I moved the project files from subfolder of the project place_pictures and this caused this software to stop working.
To be more precise I moved the folder containing this software from folder 'place_pictures' to another folder
and I made a symbolic link to this software in folder 'place_pictures/lib'
and I expected it to continue working but it did not.
I do not remember the exact error message that Python interpreter displayed
but there was something wrong with importing.

It seems I do not understand the import system in Python 3 .
I read about the import system in Python 3 related to my situation,
but I have not found an explanation that I would understand.
I came to this correction that made this software work again by trial and error.